pyproject.toml
[tool.poetry]
name = "django-tg-bot-framework"
version = "0.1.0"
description = "Framework for building telegram bots"
authors = ["Ilya Osipov <070809010@mail.ru>", "Evgeny Evseev <pelid80@gmail.com>"]
license = "GNU GPLv3"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
django = "^4"
tg_api = "1.0.0"
pytest-django = "4.5.2"
django-storages = "^1.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
